Refactor InventoryPage and inventory component exports for improved structure and maintainability
- Consolidated inventory component exports in index.js for cleaner imports.
- Updated InventoryPage to utilize new component structure, enhancing readability and organization.
- Integrated custom hook for product data management, streamlining state handling and improving performance.
