+++ src/components/inventory/index.js
@@ -1,7 +1,25 @@
-export { default as InventoryProductCard } from './InventoryProductCard';
-export { default as InventoryProductList } from './InventoryProductList';
-export { default as InventoryFilterDrawer } from './InventoryFilterDrawer';
-export { default as StockAdjustmentDialog } from './StockAdjustmentDialog';
-export { default as SupplierOrderForm } from './SupplierOrderForm';
-export { default as SupplierOrdersList } from './SupplierOrdersList';
-export { default as ProductTransactionHistory } from './ProductTransactionHistory';
+import ErrorDisplay from './ErrorDisplay';
+import InventoryFilterDrawer from './InventoryFilterDrawer';
+import InventoryHeader from './InventoryHeader';
+import InventoryProductCard from './InventoryProductCard';
+import InventoryProductContent from './InventoryProductContent';
+import InventoryProductList from './InventoryProductList';
+import InventorySearchFilters from './InventorySearchFilters';
+import ProductTransactionHistory from './ProductTransactionHistory';
+import StockAdjustmentDialog from './StockAdjustmentDialog';
+import SupplierOrderForm from './SupplierOrderForm';
+import SupplierOrdersList from './SupplierOrdersList';
+
+export {
+  ErrorDisplay,
+  InventoryFilterDrawer,
+  InventoryHeader,
+  InventoryProductCard,
+  InventoryProductContent,
+  InventoryProductList,
+  InventorySearchFilters,
+  ProductTransactionHistory,
+  StockAdjustmentDialog,
+  SupplierOrderForm,
+  SupplierOrdersList,
+};
