+++ src/hooks/index.js
@@ -1,0 +1,3 @@
+import useInventoryProducts from './useInventoryProducts';
+
+export { useInventoryProducts };
